move chunks_offsets to diskann_quantization
